Split `NorisService` into subservices for delivery methods, payments, and tax handling. Implement `NorisConnectionSubservice` for shared database connection logic. Update module structure and imports accordingly.
