+++ nest-tax-backend/src/noris/noris.module.ts
@@ -3,11 +3,21 @@ import { Module } from '@nestjs/common'
 import { PrismaModule } from '../prisma/prisma.module'
 import ThrowerErrorGuard from '../utils/guards/errors.guard'
 import { NorisService } from './noris.service'
-import {ConnectionSubservice} from "./subservices/connection.subservice";
+import { NorisConnectionSubservice } from './subservices/noris-connection.subservice'
+import { NorisDeliveryMethodSubservice } from './subservices/noris-delivery-method.subservice'
+import { NorisPaymentSubservice } from './subservices/noris-payment.subservice'
+import { NorisTaxSubservice } from './subservices/noris-tax.subservice'
 
 @Module({
   imports: [PrismaModule],
-  providers: [NorisService,ConnectionSubservice, ThrowerErrorGuard],
+  providers: [
+    NorisService,
+    NorisConnectionSubservice,
+    NorisDeliveryMethodSubservice,
+    NorisPaymentSubservice,
+    NorisTaxSubservice,
+    ThrowerErrorGuard,
+  ],
   exports: [NorisService],
   controllers: [],
 })
